refactor: Simplified BooleanField Default Value in Order Model
This commit simplifies the `completed` field within the `Order` model by removing the `null=True` attribute from the BooleanField. The `completed` field now only includes the `default=False` and `blank=True` attributes, resulting in a cleaner and more straightforward implementation.

refactor: Renamed Fields in Order and OrderItem Models

This commit addresses the renaming of specific fields within the `Order` and `OrderItem` models. The field names have been updated from their previous forms to new names that more accurately reflect their purpose and meaning. The changes encompass the following field names:

In the `Order` model:
- `get_cart_total_no_discount_past` replaces `get_cart_total_no_discount_pass`
- `get_cart_total_discount_past` replaces `get_cart_total_discount_pass`
- `get_cart_total_profit_past` replaces `get_cart_total_profit_pass`
- `get_cart_total_past` replaces `get_cart_total_pass`
- `get_cart_items_past` replaces `get_cart_items_pass`

In the `OrderItem` model:
- `get_total_no_discount_past` replaces `get_total_no_discount_pass`
- `get_total_discount_past` replaces `get_total_discount_pass`
- `get_total_profit_past` replaces `get_total_profit_pass`
- `get_total_past` replaces `get_total_pass`

Additionally, the old field names have been updated to reflect the new names throughout the codebase, ensuring consistency and clarity in the usage of these fields. This change aims to enhance the understandability of the code and improve its maintainability.