refactor: Improved Language Selection Form Styling
This commit enhances the visual styling of the language selection form for a more appealing and user-friendly appearance. The changes include:

- Adding CSS classes to the form elements for better alignment and spacing.
- Adjusting padding for better alignment.
- Utilizing a more user-friendly button style.

These improvements result in a more visually appealing language selection form that contributes to a better user experience.

refactor: Updated INSTALLED_APPS Configuration

This commit reflects an update in the `INSTALLED_APPS` configuration, replacing `'rest_framework'` with `'jalali_date'` to incorporate the `jalali_date` app into the project.

refactor: Added 'media/' to .gitignore

This commit includes the addition of the 'media/' directory to the .gitignore file. This change ensures that media files are excluded from version control, preventing them from being committed to the repository.