refactor: Updated Star Rating Display
In this commit, the HTML code for displaying star ratings has been updated. The style of stars has been modified to work dynamically with the `comment.star` variable, ensuring that the appropriate number of stars is displayed for each comment's rating. This change enhances the visual representation of ratings in the user interface.

Additionally, the corresponding styles in the main.css file have been adjusted to work seamlessly with this dynamic approach, ensuring a consistent and visually appealing presentation of comment star ratings.

refactor: Removed Unused Wishlist Icon

This commit removes an unused wishlist icon from the `_base.html` template. The icon, represented by the class "flaticon-like," is no longer required and has been safely removed from the template to simplify the codebase.

refactor: Search Query Condition for White Space Handling

In this commit, the code has been updated to handle white space more effectively. The condition `'if not request.GET.get('q'):'` has been refactored to `'if not q or q.isspace():'`. This modification ensures that the code properly detects empty or whitespace-only search queries, improving the reliability of the search functionality.

fix: Comment Queryset in ProductDetailView

This commit fixes the queryset for displaying comments in the `ProductDetailView`.

Previously, it displayed all comments with `confirmation=True` regardless of the product. Now, the queryset has been refined to only fetch comments with `confirmation=True` that are associated with the specific product being viewed (`product=self.object.pk`). This change ensures that comments are displayed correctly in the context of the product being viewed.